Move file input operation to top level of filelogreceiver
