+++ receiver/filelogreceiver/testdata/config.yaml
@@ -1,9 +1,8 @@
 receivers:
   filelog:
-    operators:
-      - type: file_input
-        include: [ receiver/filelogreceiver/testdata/simple.log ]
-        start_at: beginning
+    include: [ testdata/simple.log ]
+    start_at: beginning
+    parsers:
       - type: regex_parser
         regex: '^(?P<time>\d{4}-\d{2}-\d{2}) (?P<sev>[A-Z]*) (?P<msg>.*)$'
         timestamp:
