+++ receiver/stanzareceiver/config.go
@@ -15,26 +15,48 @@
 package stanzareceiver
 
 import (
-	"github.com/open-telemetry/opentelemetry-log-collection/pipeline"
+	"github.com/open-telemetry/opentelemetry-log-collection/operator"
+	"go.opentelemetry.io/collector/config/configmodels"
 	"gopkg.in/yaml.v2"
 )
 
-// OperatorConfig is an alias that allows for unmarshaling outside of mapstructure
+// BaseConfig is the common configuration of a stanza-based receiver
+type BaseConfig struct {
+	configmodels.ReceiverSettings `mapstructure:",squash"`
+	Parsers                       ParserConfigs `mapstructure:"parsers"`
+}
+
+// ParserConfigs is an alias that allows for unmarshaling outside of mapstructure
 // Stanza operators should will be migrated to mapstructure for greater compatibility
 // but this allows a temporary solution
-type OperatorConfig []map[string]interface{}
+type ParserConfigs []map[string]interface{}
 
-// DecodeOperators is an unmarshaling workaround for stanza operators
+// InputConfig is an alias that allows unmarshaling outside of mapstructure
+// This is meant to be used only for the input operator
+type InputConfig map[string]interface{}
+
+// decodeParserConfigs is an unmarshaling workaround for stanza operators
 // This is needed only until stanza operators are migrated to mapstructure
-func DecodeOperators(cfg OperatorConfig) (pipeline.Config, error) {
-	yamlBytes, err := yaml.Marshal(cfg)
+func (cfg BaseConfig) decodeParserConfigs() ([]operator.Config, error) {
+	if len(cfg.Parsers) == 0 {
+		return []operator.Config{}, nil
+	}
+
+	yamlBytes, err := yaml.Marshal(cfg.Parsers)
 	if err != nil {
 		return nil, err
 	}
 
-	var pipelineCfg pipeline.Config
-	if err := yaml.Unmarshal(yamlBytes, &pipelineCfg); err != nil {
+	parserCfgs := []operator.Config{}
+	if err := yaml.Unmarshal(yamlBytes, &parserCfgs); err != nil {
 		return nil, err
 	}
-	return pipelineCfg, nil
+	return parserCfgs, nil
+}
+
+// UnmarshalOperatorConfig is provided primarily for testing
+func UnmarshalOperatorConfig(operatorYaml string) (map[string]interface{}, error) {
+	operatorCfg := map[string]interface{}{}
+	err := yaml.Unmarshal([]byte(operatorYaml), &operatorCfg)
+	return operatorCfg, err
 }
